HastaneApp/requirements.txt
anyio                         4.4.0
arabic-reshaper               3.0.0
argon2-cffi                   23.1.0
argon2-cffi-bindings          21.2.0
arrow                         1.3.0
asgiref                       3.7.2
asn1crypto                    1.5.1
asttokens                     2.4.1
async-lru                     2.0.4
attrs                         23.2.0
Babel                         2.14.0
beautifulsoup4                4.12.2
bleach                        6.1.0
certifi                       2023.7.22
cffi                          1.16.0
charset-normalizer            2.0.11
click                         8.1.7
colorama                      0.4.6
comm                          0.2.1
cryptography                  41.0.7
cssselect2                    0.7.0
debugpy                       1.8.0
decorator                     5.1.1
defusedxml                    0.7.1
distlib                       0.3.8
Django                        4.1.13
django-allauth                0.62.1
django-debug-toolbar          3.6.0
django-environ                0.9.0
django-widget-tweaks          1.4.12
djangorestframework           3.13.1
djangorestframework-simplejwt 5.3.1
executing                     2.0.1
fastjsonschema                2.19.1
filelock                      3.13.1
fqdn                          1.5.1
html5lib                      1.1
idna                          3.3
ipykernel                     6.28.0
ipython                       8.20.0
isoduration                   20.11.0
jedi                          0.19.1
Jinja2                        3.1.2
json5                         0.9.14
jsonpointer                   2.4
jsonschema                    4.20.0
jsonschema-specifications     2023.12.1
jupyter_client                8.6.0
jupyter_core                  5.7.1
jupyter-events                0.9.0
jupyter-lsp                   2.2.1
jupyter_server                2.12.3
jupyter_server_terminals      0.5.1
jupyterlab                    4.0.10
jupyterlab_pygments           0.3.0
jupyterlab_server             2.25.2
lxml                          5.1.0
MarkupSafe                    2.1.3
matplotlib-inline             0.1.6
mistune                       3.0.2
nbclient                      0.9.0
nbconvert                     7.14.0
nbformat                      5.9.2
nest-asyncio                  1.5.8
notebook                      7.0.6
notebook_shim                 0.2.3
oauthlib                      3.2.2
oscrypto                      1.3.0
overrides                     7.4.0
packaging                     23.2
pandocfilters                 1.5.0
parso                         0.8.3
Pillow                        10.0.1
pip                           23.3.2
platformdirs                  2.6.2
prometheus-client             0.19.0
prompt-toolkit                3.0.43
psutil                        5.9.7
pure-eval                     0.2.2
pycparser                     2.21
Pygments                      2.17.2
pyHanko                       0.21.0
pyhanko-certvalidator         0.26.3
PyJWT                         2.8.0
PyPDF3                        1.0.6
pypng                         0.20220715.0
python-bidi                   0.4.2
python-dateutil               2.8.2
python-decouple               3.6
python-json-logger            2.0.7
python3-openid                3.2.0
pytz                          2021.3
pywin32                       306
pywinpty                      2.0.12
PyYAML                        6.0.1
pyzmq                         25.1.2
qrcode                        7.4.2
referencing                   0.32.1
reportlab                     3.6.13
requests                      2.31.0
requests-oauthlib             2.0.0
rest-framework-simplejwt      0.0.2
rfc3339-validator             0.1.4
rfc3986-validator             0.1.1
rpds-py                       0.16.2
Send2Trash                    1.8.2
setuptools                    65.5.0
six                           1.16.0
sniffio                       1.3.0
social-auth-app-django        5.4.1
social-auth-core              4.5.4
soupsieve                     2.5
sqlparse                      0.4.4
sslcommerz-lib                1.0
stack-data                    0.6.3
svglib                        1.5.1
terminado                     0.18.0
tinycss2                      1.2.1
tornado                       6.4
tqdm                          4.66.1
traitlets                     5.14.1
types-python-dateutil         2.8.19.20240106
typing_extensions             4.9.0
tzdata                        2023.4
tzlocal                       5.2
uri-template                  1.3.0
uritools                      4.0.2
urllib3                       1.26.18
virtualenv                    20.14.1
wcwidth                       0.2.13
webcolors                     1.13
webencodings                  0.5.1
websocket-client              1.7.0
whitenoise                    6.2.0
widgetsnbextension            3.5.2
xhtml2pdf                     0.2.8
Django~=5.0.6
djangorestframework~=3.15.2
pip~=24.0
xhtml2pdf~=0.2.15
environ~=1.0